book_summarizer/requirements.txt
transformers==4.41.2
torch==2.3.0
fastapi==0.111.0
uvicorn[standard]==0.30.1
sentence-transformers==3.0.1
faiss-cpu==1.8.0
pydantic==2.7.4
numpy<2.0.0